Test: ItemMainImg.vue에 관한 TC 분리
