+++ mission/tests/unit/ItemInfo/ItemInfo.spec.js
@@ -9,107 +9,6 @@ describe('ItemInfoPage', () => {
     expect(wrapper.find('#item-info-page').exists()).toBe(true);
   });
 
-  describe('Carousel', () => {
-    const wrapper = mount(ItemInfoPage);
-
-    it('hides product-img when there is no product-img', async () => {
-      await wrapper.setData({
-        itemInfo: {
-          productImg: [],
-        },
-      });
-
-      expect(wrapper.find('[data-test="product-img"]').exists()).toBeFalsy();
-    });
-
-    it('renders product-img when there is product-img', async () => {
-      await wrapper.setData({
-        itemInfo: {
-          productImg: [
-            'testImg',
-          ],
-        },
-      });
-
-      expect(wrapper.find('[data-test="product-img"]').exists()).toBeTruthy();
-    });
-
-    it('displays product-img from data', async () => {
-      await wrapper.setData({
-        itemInfo: {
-          productImg: [
-            'testImg',
-          ],
-        },
-      });
-
-      expect(wrapper.find('[data-test="product-img"]').attributes().src).toEqual('testImg');
-    });
-
-    it('renders N when there are N product-img', async () => {
-      const productImg = [
-        'testImg',
-        'testImg',
-        'testImg',
-      ];
-      await wrapper.setData({
-        itemInfo: {
-          productImg,
-        },
-      });
-
-      expect(wrapper.findAll('[data-test="product-img"]').length).toEqual(productImg.length);
-    });
-
-    describe('Slide Test', () => {
-      describe('When there is only one image,', () => {
-        beforeEach(async () => {
-          await wrapper.setData({
-            slideIndex: 0,
-            itemInfo: {
-              productImg: [
-                'testImg',
-              ],
-            },
-          });
-        });
-        it('hides prev-btn', async () => {
-          expect(wrapper.find('[data-test="prev-btn"]').exists().isVisible).toBeFalsy();
-        });
-        it('hides next-btn', async () => {
-          expect(wrapper.find('[data-test="next-btn"]').exists().isVisible).toBeFalsy();
-        });
-      });
-      describe('When there are many images', () => {
-        beforeEach(async () => {
-          await wrapper.setData({
-            slideIndex: 0,
-            itemInfo: {
-              productImg: [
-                'testImg',
-                'testImg',
-                'testImg',
-              ],
-            },
-          });
-        });
-        it('hides prev-btn when it is the first image', async () => {
-          const prevBtn = wrapper.find('[data-test="prev-btn"]');
-
-          expect(prevBtn.exists().isVisible).toBeFalsy();
-        });
-        it('hides next-btn when it is the last image', async () => {
-          const nextBtn = wrapper.find('[data-test="next-btn"]');
-
-          await nextBtn.trigger('click');
-          await nextBtn.trigger('click');
-
-          expect(nextBtn.exists().isVisible).toBeFalsy();
-        });
-      });
-    });
-  });
-
   describe('Seller Section', () => {
     const wrapper = mount(ItemInfoPage);
